seperated the system tray from the main window
