+++ commotion_client/GUI/main_window.py
@@ -23,6 +23,7 @@ from assets import commotion_assets_rc
 from GUI.menu_bar import MenuBar
 from GUI.crash_report import CrashReport
 
+
 class MainWindow(QtGui.QMainWindow):
     """
     The central widget for the commotion client. This widget initalizes all other sub-widgets and extensions as well as defines the paramiters of the main GUI container.
@@ -100,13 +101,6 @@ class MainWindow(QtGui.QMainWindow):
         #Create slot to monitor when menu-bar wants the main window to change the main-viewport
         self.connect(self.menu_bar, QtCore.SIGNAL("viewportRequested()"), self.changeViewport)
         
-        #Create tray
-        self.tray = TrayIcon(self)
-        #connect to tray events for closing application and showing main window.
-        self.connect(self.tray.exit, QtCore.SIGNAL("triggered()"), self.exitEvent)
-        self.connect(self.tray, QtCore.SIGNAL("showMainWindow"), self.bringFront)
-        #crash("YOU NEED TO IMPLEMENT THE CRASHING FUNCTIONALITY HERE")
-
     def toggle_menu_bar(self):
         #if menu shown... then
         #DockToHide = self.findChild(name="MenuBarDock")
@@ -209,31 +203,3 @@ class MainWindow(QtGui.QMainWindow):
             self.exitOnClose = True
             self.close()
         
-
-class TrayIcon(QtGui.QWidget):
-    """
-    The Commotion tray icon. This icon object is the only object that can close the entire application.
-    """
-
-    def __init__(self, parent=None):
-        super().__init__()
-        
-        #Create actions for tray menu
-        self.exit = QtGui.QAction(QtGui.QIcon(), "Exit", self)
-
-        #set tray Icon and it's menu which allows closing from it.
-        self.tray_icon = QtGui.QSystemTrayIcon(QtGui.QIcon(":logo32.png"), self)
-        menu = QtGui.QMenu(self)
-        menu.addAction(self.exit) #add exit action to tray icon
-        self.tray_icon.setContextMenu(menu)
-        self.tray_icon.activated.connect(self.tray_iconActivated)
-        self.tray_icon.show()
-
-    def tray_iconActivated(self, reason):
-        """
-        Defines the tray icon behavior on different types of interactions.
-        """
-        if reason == QtGui.QSystemTrayIcon.Context:
-            self.tray_icon.contextMenu().show()
-        elif reason == QtGui.QSystemTrayIcon.Trigger:
-            self.emit(QtCore.SIGNAL("showMainWindow"))
